+++ src/jaicore/search/algorithms/standard/bestfirst/RandomizedDepthFirstEvaluator.java
@@ -5,7 +5,7 @@ import java.util.Random;
 import jaicore.search.algorithms.parallel.parallelexploration.distributed.interfaces.SerializableNodeEvaluator;
 import jaicore.search.structure.core.Node;
 
-public class RandomizedDepthFirstEvaluator<T> implements SerializableNodeEvaluator<T,Integer> {
+public class RandomizedDepthFirstEvaluator<T> implements SerializableNodeEvaluator<T,Double> {
 
 	private final Random rand;
 
@@ -15,7 +15,7 @@ public class RandomizedDepthFirstEvaluator<T> implements SerializableNodeEvaluat
 	}
 
 	@Override
-	public Integer f(Node<T,Integer> node) {
-		return (int) (-1 * (node.path().size() * 1000 + rand.nextInt(100)));
+	public Double f(Node<T,Double> node) {
+		return (double) (-1 * (node.path().size() * 1000 + rand.nextInt(100)));
 	}
 }
